MDL-34356 JavaScript: Allow chooser dialogues to submit values for hidden elements
When we submit the chooser dialogue, we disable the radio buttons used for
selection. Since browsers do not actually submit values for disabled
elements, we store the radio value in a hidden element which is not
disabled.

This commit abstracts this out of the activity chooser and into the generic
chooser dialogue such that other types of chooser dialogue may use it.