MDL-42889 cleanup obsolete clam relatd code
Since 2.0 all uploaded files need to go through the file picker which does the scanning,
the idea is that it is more important to warn users that their computer is crawling
with viruses. All teachers are supposed to have local up-to-date antivirus software!

Since 2.0 we are not scanning existing files on server, in future we might use the
status flag in files table to indicate the file is problematic, but still we should
never detele anthing from the filedir...