MDL-43856 MathJax: alter the equation editor to make it use MathJax if it is available.
If MathJax is loaded by any means, the editor should use it instead of the filter. This prevents
slow ajax requests and gives much better rendering. MathJax could be loaded by the filter,
or it could just be done the old way by sticking the include in the page header.