MDL-47046 gradereport_grader: Update JS to meet coding style guidelines
This also moves the code into a set of separate files to allow for
additional components of the report to be migrated to the same YUI module,
adjusts some UI components, improves JS performance, and adds appropriate
documentation.

This issue is a part of the MDL-46658 Task.
This issue is a part of the MDL-25544 Epic.