MDL-48070 navigation: optimised the login page
The navigation and settings no longer get initialised on the login
page by default, this is achieved by calling to ignore the active
page on the navbar.
It is safe because the login page is not represented in either the
navigation or the settings.

This change saves a few MB or ram and a couple of database queries