MDL-46531 editor_atto: Handle a null selection better on colour change
Sadly, there is no easy way to do this. When a range is empty, we need to
create a new span, make the changes there, and then select it. Rangy does
not provide a helper to do this yet (will be in 1.3) so we have to do it
manually in the mean time.