MDL-48177 core: Plugin environment checks with no version of Moodle
The environment.xml files in plugins force you to define the versions of Moodle that the checks should be carried out against.
For community plugins the environment requirements of the plugin seem unlikely to be dependant on the version of Moodle.

This patch will allow the environment.xml files to use a new <PLUGIN> tag in place of the <MOODLE version="> tag.

This solution has the following properties:

* The Environment file will not break versions of Moodle that do not support the PLUGIN syntax.
* Specific versioned MOODLE tags can still be defined for versions of Moodle that do not have this patch.
* If both MOODLE and PLUGIN tags exist the PLUGIN tag will be used in preference.
* The PLUGIN tag will not function in the main Moodle environment.xml file.
* The PLUGIN tag has a mandatory name attribute that must match the frankenstyle name of the plugin being tested.