MDL-48898 quiz: use admin_setting_configduration
Now this setting type is available, it is a better UI.

This required a new setting class
admin_setting_configduration_with_advanced.

Also, I took the liberty of shortening the default auto-save delay to 1
minute, and re-wording the description of that setting, now that it can
be changed to any value.