MDL-49356 restore: handle course completion if still not defined
Before this patch course completion only was restored if, appart from
other conditions... we were restoring to new course, never to existing
one.

This relaxes a bit that condition in order to be able to restore to an
existing course not having any course completion setting defined, so no
conflict will happen ever.