MDL-50342: Convert awarding of badges to be event based.
This patch switches Moodle's issuing of badges from being cron based to event
based. This avoids unnecessary and inefficient checking and rechecking of
criteria that haven't yet been fulfilled when nothing has changed since the
last cron run.

The patch has been made as simple as possible. It checks the criteria for all
unissued badges for a user when an event occurs that could result in criteria
being fulfilled. Further optimisations could potentially provide the ability to
just check the criteria (if any) related to the event.

This version of the patch will allow a badge to be issued prior to the start
date for a course. It could be modified further so that Moodle didn't display
the badge or send a notification until the start date, but I've left that
out for now.

Signed-off-by: Nigel Cunningham <nigelc@catalyst-au.net>