MDL-44273 block_rss: Back-off strategy for RSS feeds
Based on a patch by Davo Smith of Synergy Learning

RSS feeds can fail, either permanently or temporarily.
After this patch if a feed fails it will not be tried again in the next cron run.
If it continues to fail the delay between tries will be increased.
When a feed succeeds it will go back to being fetched on each cron run.

This should make the CRON process for RSS feeds finish faster when
there are lots of invalid feeds.